main.c/h  added map initialization and render  added monster initialization and render  added a rat to the monsters  added some new headers player.c/h  moved the player instance here  changed some function to use the local player instance  modified the implementation of playerHandleInput to confine the player   in the map and permit movement only on walkable tiles  added playerRender  added some header screen.c/h  changed screenWidth and screenHeight type signal.c/h  changed sigHandler to signalHandler map.c/h  this files will contain all map related functions and data types  added a mapType_t to have different type of map  added a map_t that represents a map  added mapInit to initialize the map based on its type  added mapIsValid to check for out of bounds coordinates  added mapRender to render the map  added mapTileAt to retrive tiles from the map mapSample.c/h  this files will contain all MAP_SAMPLE map type related functions monster.c/h  this files will contain all (single) monster related functions and data   types  added monsterType_t to have different type of monster  added monster_t that represents a monster  added monsterInit that initialize a monster based on its type  added monsterGlyph to retrive a monster type character representation  added monsterColor to retrive a mosnter type color monsters.c/h  this files will contain all monsters related functions and data types  added monsters which is an array of actual existing monsters  added monstersInit which initialize the monsters array  added monstersAdd which add a monster to the array based in its type  added monstersRender which draw all mosnters tile.c/h  this files will contain all tile related functions and data types  added tileType_t to have different tiles  added tile_t that represents a tile  added tileInit that initialize a tile based on its type  added tileGliph to retrive a tile type character representation  added tileFGColor to retrive a tile type foreground color  added tileBGColor to retrive a tile type background color  added tileWalkable to retrive the walkable flag of a tile type  added tileBlockFOV to retrive the block field of view flag of a tile   type

+++ screen.c
@@ -5,8 +5,8 @@
 //https://en.wikipedia.org/wiki/ANSI_escape_code
 
 //These two varaibles will hold the terminal size after a call to screenInit
-int screenWidth=0;
-int screenHeight=0;
+uint16_t screenWidth=0;
+uint16_t screenHeight=0;
 
 //Initialize the screen: get terminal size and hide the cursor
 void screenInit(void)
